% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APJaccard.R
\name{FindJaccardMatrix}
\alias{FindJaccardMatrix}
\title{A function that calculates a trnasformed Jaccard similarity matrix for the cells in a Seurat object}
\usage{
FindJaccardMatrix(object, k = 5, transformation = "default")
}
\arguments{
\item{object}{a Seurat object which has had a pca reduction performed on it}

\item{k}{an integer representing the number of neighbors which will be found for each cell}

\item{transformation}{a character vector indicating the type of transformation the similarity matrix will go through before being returned. By default the similairty matrix will be returned with no transformation. Other possible values include log (returns matrix which has been log transformed), 'inverse' returns a shifted inverse of the matrix, 'stretch' returns a linear scaling of the matrix, 'tan' returns the tangent of matrix values, and 'tanh' returns the hyperbolic tangent of matrix values}
}
\value{
an nxn matrix where n is the number of cells in the Seurat object and the matrix contians the transformed Jaccard values for all pairs of cells
}
\description{
A function that calculates a trnasformed Jaccard similarity matrix for the cells in a Seurat object
}
