% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APJaccard.R
\name{FindNMI}
\alias{FindNMI}
\title{A function that computes the Normalized Mutual Information  (NMI) between two 
cluster vectors.}
\usage{
FindNMI(clusters1, clusters2)
}
\arguments{
\item{clusters1}{a (named) vector with elements being assignment labels}

\item{clusters2}{a (named) vector with elements being assignment labels}
}
\value{
A numerical value on the range [0,1] corresponding to the NMI
}
\description{
A function that computes the Normalized Mutual Information  (NMI) between two 
cluster vectors.
}
