% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APJaccard.R
\name{plotClusters}
\alias{plotClusters}
\title{A function that creates 2-D plot of single-cells, colored by any clustering solution}
\usage{
plotClusters(object, clusters, plotName)
}
\arguments{
\item{object}{a Seurat object which has had a pca, umap, or tsne reduction performed on it}

\item{clusters}{a vector of cluster assignments, labeled with cell names}

\item{plotName}{a ggplot object with cells plotted by the dimensional reduction of the object and colors determined by the assignment from clusters}
}
\value{
....
}
\description{
A function that creates 2-D plot of single-cells, colored by any clustering solution
}
