% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APJaccard.R
\name{scoreAPTest}
\alias{scoreAPTest}
\title{A function that calculates the NMI values for all vectors in a bigresult 
object (returned from APTest function) as compared with one vector.}
\usage{
scoreAPTest(bigresult, assignments, SHUTUP = TRUE)
}
\arguments{
\item{bigresult}{an object with the clustering solutions of all iterations of the APTest algorithm, returned from APTest}

\item{assignments}{a vector with the cluster assignments for all cells, named by the cell names, containing the same number of cells as bigresult}
}
\value{
a vector of NMI values and their corresponding APTest results
}
\description{
A function that calculates the NMI values for all vectors in a bigresult 
object (returned from APTest function) as compared with one vector.
}
